minor: refactor to move some shuffle related logic from QueryPlanSerde to CometExecRule
